기본적인 서버 api 기능 전체 완성
* Member API 완성
* Library API 완성

ref #11
