+++ MYLIB_SRC/mylib/src/main/java/com/js/mylib/controller/LibraryApiController.java
@@ -1,8 +1,6 @@
 package com.js.mylib.controller;
 
-import com.js.mylib.dto.CreateLibraryRequestDto;
-import com.js.mylib.dto.LibraryInfoDto;
-import com.js.mylib.dto.LibrarySearchConditionDto;
+import com.js.mylib.dto.*;
 import com.js.mylib.service.LibraryService;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +9,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -28,14 +27,38 @@ public class LibraryApiController {
         return libraryService.findLibraryInfoById(libraryId);
     }
 
+    @GetMapping("/api/v1/libraries/find/library/member/{libraryId}")
+    public List<MemberInfoDto> findLibraryMemberByLibraryId(@PathVariable Long libraryId) {
+        return libraryService.findLibraryMemberByLibraryId(libraryId);
+    }
+
     @GetMapping("/api/v1/libraries")
     public Page<LibraryInfoDto> searchLibraries(LibrarySearchConditionDto libraryCondition, Pageable pageable) {
         return libraryService.searchLibraries(libraryCondition, pageable);
     }
 
+    @PutMapping("/api/v1/libraries/update/library/{libraryId}/member/{memberId}")
+    public LibraryResponse updateLibrary(@PathVariable Long libraryId, @PathVariable Long memberId, @RequestBody @Valid UpdateLibraryRequestDto updateLibraryRequest) {
+        Long id = libraryService.updateLibrary(libraryId, memberId, updateLibraryRequest);
+        return new LibraryResponse(id);
+    }
+
+    @DeleteMapping("/api/v1/libraries/exit/library/{libraryId}/member/{memberId}")
+    public LibraryResponse exitLibrary(@PathVariable Long libraryId, @PathVariable Long memberId) {
+        Long id = libraryService.exitLibrary(memberId, libraryId);
+        return new LibraryResponse(id);
+    }
+
+    @PostMapping("/api/v1/libraries/enter/library/{libraryId}/member/{memberId}")
+    public LibraryResponse enterLibrary(@PathVariable Long libraryId, @PathVariable Long memberId) {
+        Long id = libraryService.enterLibrary(libraryId, memberId);
+        return new LibraryResponse(id);
+    }
+
     @Data
     static class LibraryResponse {
         private Long id;
+        private String message;
         public LibraryResponse(Long id) {
             this.id = id;
         }
