+++ MYLIB_SRC/mylib/src/main/java/com/js/mylib/dto/UpdateLibraryRequestDto.java
@@ -1,0 +1,11 @@
+package com.js.mylib.dto;
+
+import lombok.Data;
+
+@Data
+public class UpdateLibraryRequestDto {
+	private String name;
+	private String description;
+	private String endTime;
+	private int memberLimit;
+}
