+++ MYLIB_SRC/mylib/src/main/java/com/js/mylib/service/MemberService.java
@@ -6,14 +6,12 @@ import com.js.mylib.dto.UpdateMemberRequestDto;
 import com.js.mylib.entity.Member;
 import com.js.mylib.entity.MemberType;
 import com.js.mylib.dto.JoinMemberRequestDto;
-import com.js.mylib.repository.LibraryJPARepository;
-import com.js.mylib.repository.MemberJPARepository;
+import com.js.mylib.repository.CustomQueryRepository;
 import com.js.mylib.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -22,9 +20,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class MemberService {
 	private final MemberRepository memberRepository;
-	private final LibraryService libraryService;
-	private final MemberJPARepository memberJPARepository;
-	private final LibraryJPARepository libraryJPARepository;
+	private final CustomQueryRepository customQueryRepository;
 	public List<MemberInfoDto> findMembers() {
 		return memberRepository.findAllMember();
 	}
@@ -39,14 +35,6 @@ public class MemberService {
 
 	}
 
-	public List<MemberInfoDto> findLibraryMemberByLibraryId(Long libraryId) {
-		List<Member> LibraryMembers = libraryService.findLibraryById(libraryId).getMembers();
-		List<MemberInfoDto> result = LibraryMembers.stream()
-				.map(m -> new MemberInfoDto(m))
-				.collect(Collectors.toList());
-		return result;
-	}
-
 	public Long saveMember(JoinMemberRequestDto request) {
 		MemberType memberType = null;
 		if(request.getType().equals("학생")) {
@@ -59,7 +47,7 @@ public class MemberService {
 			memberType = MemberType.ETC;
 		}
 		Member member = new Member(request.getName(), request.getEmail(), request.getPassword(), memberType);
-		memberJPARepository.saveMember(member);
+		customQueryRepository.saveMember(member);
 		return member.getId();
 	}
 
@@ -79,37 +67,46 @@ public class MemberService {
 			memberType = MemberType.ETC;
 		}
 
-		memberJPARepository.updateMember(memberId, request, memberType);
+		customQueryRepository.updateMember(memberId, request, memberType);
 		return memberId;
 
 	}
 
-	public Long enterLibrary(Long libraryId, Long memberId) {
-		return memberJPARepository.enterLibrary(libraryId, memberId);
-
-	}
-
-	@Transactional
-	public Long exitLibrary(Long memberId, Long libraryId) {
-
-		LibraryInfoDto libraryInfo = libraryService.findLibraryInfoById(libraryId);
-		List<MemberInfoDto> libraryMember = findLibraryMemberByLibraryId(libraryId);
-
-		// 남은 회원이 한명일 때는 삭제 도서관 삭제
-		if(libraryMember.size() == 1) {
-			if(memberId == libraryInfo.getReaderId()) {
-				libraryService.deleteLibrary(libraryId);
-			}
-		} else { //회원이 여러명인데 리더가 아니면
-			//리더가 아니면
-			if(memberId != libraryInfo.getReaderId()) {
-				libraryService.exitLibrary( memberId, libraryId);
-			} else { //리더라면 나가면서 리더 교체
-				System.out.println("들어옴??");
-				libraryService.changeReader(memberId, libraryId);
-			}
-		}
-
-		return memberId;
-	}
+//	public Long enterLibrary(Long libraryId, Long memberId) {
+//		return customQueryRepository.enterLibrary(libraryId, memberId);
+//
+//	}
+//
+//	@Transactional
+//	public Long exitLibrary(Long memberId, Long libraryId) {
+//
+//		LibraryInfoDto libraryInfo = libraryService.findLibraryInfoById(libraryId);
+//		List<MemberInfoDto> libraryMember = findLibraryMemberByLibraryId(libraryId);
+//
+//		if(libraryInfo == null) return -1l;
+//		boolean isExistMember = false;
+//		for(int i =0; i<libraryMember.size(); i++) {
+//			if(libraryMember.get(i).getMemberId() == memberId ) {
+//				isExistMember = true;
+//			}
+//		}
+//
+//		if(!isExistMember) return -1l;
+//
+//		// 남은 회원이 한명일 때는 삭제 도서관 삭제
+//		if(libraryMember.size() == 1) {
+//			if(memberId == libraryInfo.getReaderId()) {
+//				libraryService.deleteLibrary(memberId,libraryId);
+//			}
+//		} else { //회원이 여러명인데 리더가 아니면
+//			//리더가 아니면
+//			if(memberId != libraryInfo.getReaderId()) {
+//				libraryService.exitLibrary( memberId, libraryId);
+//			} else { //리더라면 나가면서 리더 교체
+//				libraryService.changeReader(memberId, libraryId);
+//			}
+//		}
+//
+//		return memberId;
+//	}
 }
